fix db module

+++ src/db/types.rs
@@ -1,8 +1,6 @@
-use chrono::NaiveDateTime;
 use serde::{Deserialize, Serialize};
-use sqlx::types::BigDecimal;
-
-#[derive(Debug, Clone, sqlx::FromRow, Serialize, Deserialize)]
+use utoipa::ToSchema;
+#[derive(Debug, Clone, sqlx::FromRow, Serialize, Deserialize, ToSchema)]
 pub struct Proposal {
     pub id: i32,
     pub author_id: String,
@@ -12,7 +10,7 @@ pub struct Proposal {
 pub struct ProposalSnapshot {
     pub proposal_id: i32,
     pub block_height: i64,
-    pub ts: BigDecimal,
+    pub ts: i32,
     pub editor_id: String,
     pub social_db_post_block_height: i64,
     pub labels: serde_json::Value,
@@ -24,7 +22,7 @@ pub struct ProposalSnapshot {
     pub description: Option<String>,
     pub linked_proposals: Option<serde_json::Value>,
     pub linked_rfp: Option<i32>,
-    pub requested_sponsorship_usd_amount: Option<BigDecimal>,
+    pub requested_sponsorship_usd_amount: Option<i32>,
     pub requested_sponsorship_paid_in_currency: Option<String>,
     pub requested_sponsor: Option<String>,
     pub receiver_account: Option<String>,
@@ -38,7 +36,7 @@ pub struct Dump {
     pub receipt_id: String,
     pub method_name: String,
     pub block_height: i64,
-    pub block_timestamp: BigDecimal,
+    pub block_timestamp: i32,
     pub args: String,
     pub author: String,
     pub proposal_id: i64,
@@ -49,7 +47,7 @@ pub struct ProposalWithLatestSnapshot {
     pub proposal_id: i32,
     pub author_id: String,
     pub block_height: i64,
-    pub ts: BigDecimal,
+    pub ts: i32,
     pub editor_id: String,
     pub social_db_post_block_height: i64,
     pub labels: serde_json::Value,
@@ -61,7 +59,7 @@ pub struct ProposalWithLatestSnapshot {
     pub description: Option<String>,
     pub linked_proposals: Option<serde_json::Value>,
     pub linked_rfp: Option<i32>,
-    pub requested_sponsorship_usd_amount: Option<BigDecimal>,
+    pub requested_sponsorship_usd_amount: Option<i32>,
     pub requested_sponsorship_paid_in_currency: Option<String>,
     pub requested_sponsor: Option<String>,
     pub receiver_account: Option<String>,
@@ -80,7 +78,7 @@ pub struct Rfp {
 pub struct RfpSnapshot {
     pub rfp_id: i32,
     pub block_height: i64,
-    pub ts: BigDecimal,
+    pub ts: i32,
     pub editor_id: String,
     pub social_db_post_block_height: i64,
     pub labels: serde_json::Value,
@@ -92,7 +90,7 @@ pub struct RfpSnapshot {
     pub summary: Option<String>,
     pub description: Option<String>,
     pub timeline: Option<serde_json::Value>,
-    pub submission_deadline: BigDecimal,
+    pub submission_deadline: i32,
     pub views: Option<i32>,
 }
 
@@ -101,7 +99,7 @@ pub struct RfpWithLatestSnapshot {
     pub rfp_id: i32,
     pub author_id: String,
     pub block_height: i64,
-    pub ts: BigDecimal,
+    pub ts: i32,
     pub editor_id: String,
     pub social_db_post_block_height: i64,
     pub labels: serde_json::Value,
@@ -114,7 +112,7 @@ pub struct RfpWithLatestSnapshot {
     pub description: Option<String>,
     pub timeline: Option<serde_json::Value>,
     pub views: Option<i32>,
-    pub submission_deadline: BigDecimal,
+    pub submission_deadline: i32,
 }
 
 #[derive(Debug, Clone, sqlx::FromRow, Serialize, Deserialize)]
@@ -122,7 +120,7 @@ pub struct RfpDump {
     pub receipt_id: String,
     pub method_name: String,
     pub block_height: i64,
-    pub block_timestamp: BigDecimal,
+    pub block_timestamp: i32,
     pub args: String,
     pub author: String,
     pub rfp_id: i64,
